feat(lib): add query batches navigation drawer
